refactor: 퇴원/면회 차트 폴더 구조 변경
