+++ components/hospital/icu/main/chart/selected-chart/chart-body/table/chart-table.tsx
@@ -17,7 +17,7 @@ import { useIcuOrderStore } from '@/lib/store/icu/icu-order'
 import { formatOrders, sortOrders } from '@/lib/utils'
 import { useBasicHosDataContext } from '@/providers/basic-hos-data-context-provider'
 import type { SelectedChart, SelectedIcuOrder } from '@/types/icu/chart'
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import { useDebouncedCallback } from 'use-debounce'
 import CellsRow from './cells-row'
 import CellsRowTitle from './cells-row-title'
