+++ components/hospital/icu/main/out-and-visit/out-and-visit-skeleton.tsx
@@ -1,0 +1,6 @@
+import { Skeleton } from '@/components/ui/skeleton'
+import React from 'react'
+
+export default function OutAndVisitSkeleton() {
+  return <Skeleton className="h-[400px] w-full" />
+}
